Fixed a bunch of null pointer exceptions, changed id's to long's, fixed formatting on insertion statements, fixed sql statements to match documentation names, and a few other things.

+++ src/main/java/com/rainforestcommerce/rcdb/models/Shipment.java
@@ -7,17 +7,17 @@ import java.util.Date;
  */
 public class Shipment {
 
-    Integer id;
+    long id;
     String store;
     Date requestDate;
 
-    public Shipment(Integer ID, String store, Date requestDate){
+    public Shipment(long ID, String store, Date requestDate){
         this.id = ID;
         this.store = store;
         this.requestDate = requestDate;
     }
 
-    public Integer getID() {
+    public long getID() {
         return id;
     }
